try_lance/pyproject.toml
[project]
name = "try-lance"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "t_w", email = "tzmtwtr@yahoo.co.jp" }
]
dependencies = [
    "pylance>=0.9.17",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "duckdb>=0.10.0",
    "numpy>=1.26.4",
    "tqdm>=4.66.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "try_lance:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/try_lance"]
